Fixes problems for short captions that arise in 2 conditions:
(1) when figure title is empty -- when we have a figure of
the form:

    [Caption](image.jpg){#fig:ref}

In this case, the earlier code produces:

    \caption[]{Caption}

and the captions package would then print out the '[]'. This avoids
that.

(2) when the figure is unnumbered: then having a short
caption is senseless. The earlier code would produce:

    \caption*[short caption]{Caption}

and the captions package would include the '[short caption'] along with
the caption. Again, this avoids this problem.