default id for numbered figures with no id
Previous behaviour was to assign these figures an id of '', which meant
that all figures without an id ended up with the same figure number
(equal to the total number of figures).

Now figures without an id will be given a unique one that is derived
from the figure count at the time. The id can be recovered when
replacing figures by maintaining another count for replaced figures.