Fix citation prefix for internal references
Currently, `pandoc-reference-filter` creates an internal reference by
adding together the citation prefix, a space, the link, and the
citation suffix. This can cause problems when the reference is of the
form:

    (@sec:section-name)

In this case, there is no citation prefix, but a space is inserted
anyway, resulting in output of the form:

    ( Section 1)

This removes the initial space when there is no citation prefix.