Prevent adding extra space to bibliographic references
If a reference is to an internal object and has a prefix, there needs to
be a space inserted between the prefix and the generated reference.
However, that space cannot be added when the reference is a
bibliographic reference. This fix moves the addition of a space to after
the check for bibliographic reference.